Form model (#1990)
* Create new FormModel

Merges responsibility for enrichForm, progress, and validateForm into a
single combined model.

* Move locationsFor to awards-for-all

* Use getCurrentFieldsForStep in step router

* Use getCurrentFieldsForStep in tests, remove fieldsFor

* Update pagination to accept sectionSlug

* Inline pagination into form model

* Remove validateForm in favour of form.validate

* Inline mocks into form.test.js

* Fixup tests
